Continued error handling for Post and Media. Configured Get Posts by User view

+++ backend/surreel/api/views/posts.py
@@ -3,23 +3,23 @@ from rest_framework.decorators import api_view
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework.parsers import JSONParser
 from rest_framework.serializers import ValidationError
-from ..utils.helper_functions import format_validation_errors
+from rest_framework.response import Response
+from rest_framework import status
+from django.shortcuts import get_object_or_404
 from ..models import Post
 from ..serializers import PostSerializer
 from .media import create_media
-from ..exceptions.media import NoMediaForPost
 
 
 
 # View Functions #
 
 
-'''  /  '''
+''' / '''
 
 @api_view(['GET', 'POST'])
 @csrf_exempt
 def post_list(request):
-
     if request.method == 'GET':
         return get_all_posts()
 
@@ -27,16 +27,12 @@ def post_list(request):
         return create_post(request)
 
 
-'''  /<id>  '''
+''' /<post_id> '''
 
 @api_view(['GET', 'PATCH', 'DELETE'])
 @csrf_exempt
 def post_details(request, pk):
-
-    try:
-        post = Post.objects.get(id=pk)
-    except Post.DoesNotExist:
-        return JsonResponse({"error": "Post not found", "status": 404})
+    post = get_object_or_404(Post, pk=pk)
 
     if request.method == 'GET':
        return get_post_by_id(post)
@@ -48,70 +44,62 @@ def post_details(request, pk):
         return delete_post(post)
 
 
+''' /<user_id>/posts '''
+
+@api_view(['GET'])
+@csrf_exempt
+def user_posts(request, pk):
+    posts = Post.objects.filter(user=pk)
+    serializer = PostSerializer(posts, many=True)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+
 # Helper Functions #
 
 def get_all_posts():
     posts = Post.objects.all()
     serializer = PostSerializer(posts, many=True)
-    data_dict = {'posts': serializer.data}
-
-    return JsonResponse(data_dict)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 def get_post_by_id(post):
-    post_serializer = PostSerializer(post)
-    return JsonResponse(post_serializer.data)
+    serializer = PostSerializer(post)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 def create_post(request):
-    try:
-        post_data = JSONParser().parse(request)
-        media_data = post_data.pop('media', None)
+    media_data = request.data.pop('media', None)
 
-        if media_data == None:
-            raise NoMediaForPost()
+    if media_data == None or not media_data:
+        return Response({"detail" : "You must provide at least one media file for your post."}, status=status.HTTP_404_NOT_FOUND)
 
-        post_serializer = PostSerializer(data=post_data)
+    post_serializer = PostSerializer(data=request.data)
 
-        if post_serializer.is_valid():
-            post_instance = post_serializer.save()
-            post_id = post_instance.id
-        else:
-            raise ValidationError(post_serializer.errors)
+    if post_serializer.is_valid():
+        post_instance = post_serializer.save()
 
-        media_error = create_media(media_data, post_id)
+        media_creation_errors = create_media(media_data, post_instance.id)
 
-        if media_error is not None:
+        if media_creation_errors is not None:
             post_instance.delete()
-            return media_error
+            return Response(media_creation_errors, status=status.HTTP_400_BAD_REQUEST)
 
-        return  JsonResponse(post_serializer.data, status=201)
+        return Response(post_serializer.data, status=status.HTTP_201_CREATED)
 
-    except NoMediaForPost as e:
-        return JsonResponse({"error": str(e), "status" : e.status_code})
-
-    except ValidationError as e:
-        formatted_error = format_validation_errors(e)
-        return JsonResponse({"errors": formatted_error, "status": 400})
+    return Response(post_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 def delete_post(post):
     post.delete()
-    return JsonResponse({"message": "The post has been deleted", "status": 200})
+    return JsonResponse({"detail": "The post has been deleted"}, status=status.HTTP_200_OK)
 
 
 def edit_post(request, post):
-    try:
-        data = JSONParser().parse(request)
-        post_serializer = PostSerializer(post, data=data, partial=True)
+    data = JSONParser().parse(request)
+    post_serializer = PostSerializer(post, data=data, partial=True)
 
-        if post_serializer.is_valid():
-            post_serializer.save()
-            return  JsonResponse(post_serializer.data, status=200)
+    if post_serializer.is_valid():
+        post_serializer.save()
+        return Response(post_serializer.data, status=status.HTTP_200_OK)
 
-        else:
-            raise ValidationError(post_serializer.errors)
-
-    except ValidationError as e:
-        formatted_error = format_validation_errors(e)
-        return JsonResponse({"errors": formatted_error, "status": 400})
+    return Response(post_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
